requirements.txt
# Core Framework
streamlit>=1.28.0

# AI/ML Libraries
google-generativeai>=0.3.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0
torch>=2.0.0

# Document Processing
PyPDF2>=3.0.0
python-docx>=0.8.11
lxml>=4.9.0
chardet>=5.0.0
pyyaml>=6.0

# Data Processing
pandas>=2.0.0
openpyxl>=3.1.0

# API Calls
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
uuid>=1.30

# Database
pymongo>=4.5.0
pymongo[srv]>=4.5.0

# Authentication
bcrypt>=4.0.0
cryptography>=41.0.0

# Optional (for development)
# pytest>=7.0.0
# black>=23.0.0
# pylint>=2.17.0